Add files via upload
Clustering is the task of dividing the data points into a number of groups such that data points in the same groups are more similar to other data points in the same group than those in other groups. In simple words, the aim is to segregate groups with similar traits and assign them into clusters.
Types of Clustering : 
1) K-means Clustering
2) Hierarchical Clustering
3)Apritori Algorithm
Source Codes are provided in both Python and R-Programming languages.